Add Rotaion property to screen_graphic
