+++ luda-editor/rpc/src/data/cut/previous/v2.rs
@@ -1,8 +1,8 @@
 use super::*;
-use crate::data::ScreenImage;
+use crate::data::screen_graphic;
 use crate::Uuid;
 
-pub type ScreenImages = [Option<ScreenImage>; 5];
+pub type ScreenImages = [Option<screen_graphic::previous::v0::ScreenImage>; 5];
 
 #[migration::version(2)]
 #[derive(Debug, Clone, Default)]
@@ -21,7 +21,7 @@ impl Cut {
             line: previous.line,
             character_name: "".to_string(),
             screen_images: previous.screen_images.map(|image| {
-                image.map(|image| ScreenImage {
+                image.map(|image| screen_graphic::previous::v0::ScreenImage {
                     id: image.id,
                     circumscribed: image.circumscribed,
                 })
