+++ luda-editor/rpc/src/data/screen_graphic/mod.rs
@@ -1,0 +1,4 @@
+pub mod previous;
+mod v1;
+
+pub use v1::{ScreenCg, ScreenGraphic, ScreenImage};
